apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "1.0.0"
description = "VinoBuzz AI Sommelier API"
authors = ["cording12 <joncording12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
supabase = "^2.5.0"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
langchain-core = "^0.2.38"
openai = "^1.45.0"
redis = {extras = ["hiredis"], version = "^5.0.8"}

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.6.0"
pytest = "^8.0.1"
pylint-pydantic = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
